requirements.txt
qrcode==7.3.1
Pillow==8.4.0